+++ v0_021/main.py
@@ -20,7 +20,9 @@ class Game:
         fonts_folder = path.join(game_folder, 'fonts')
         # -- load fonts -- 
         self.FONT_TWINMARKER_26 = pg.font.Font((path.join(fonts_folder, "TwinMarker.ttf")), 26) 
+        self.FONT_VETERAN_TYPEWRITER_20 = pg.font.Font((path.join(fonts_folder, "veteran typewriter.ttf")), 20) 
         self.FONT_VETERAN_TYPEWRITER_26 = pg.font.Font((path.join(fonts_folder, "veteran typewriter.ttf")), 26) 
+        self.FONT_BOHEMIAN_TYPEWRITER_20 = pg.font.Font((path.join(fonts_folder, "Bohemian Typewriter.ttf")), 20)
         self.FONT_BOHEMIAN_TYPEWRITER_26 = pg.font.Font((path.join(fonts_folder, "Bohemian Typewriter.ttf")), 26)
 
     def new_level(self):
@@ -42,6 +44,7 @@ class Game:
         #
         self.all_shelved_chat_customers = []
         self.all_open_chat_customers = []
+        self.mouse_click_down = False
 
     def run(self):
         self.playing = True
@@ -65,19 +68,23 @@ class Game:
                     if clicked_shelved_customer not in self.all_open_chat_customers:
                        self.all_open_chat_customers.append(clicked_shelved_customer)
                 #
+
+                # so here we'll do a held down check that will move and also break (as we defo dont wanna do the below check (tho it shouldnt validate on down anyways duh))
+                a_customer.move_click_hold_opened_chatbox()
+
                 clicked_open_customer = a_customer.check_click_opened_chatbox()
                 if clicked_open_customer:
-                    print(f"{clicked_open_customer.my_name = }")
+                    # print(f"{clicked_open_customer.my_name = }")
                     if self.display_tabs[0] == "new_orders":   
                         self.new_orders_tab.image = self.new_orders_tab.blank_img_surf
-                        # break < adding a break means the bottom one comes off first so would be easy to figure out the inverse
-                        # to stop the jittery blit 
-                        # you just need to reblit here - or just in refactor tomo figure it out more thoroughly for the order of operations!
+                        break # break the loop on the first rect the mouse collides with as its the one on top and thats the one we want to bring to the front                         
                 else:
                     a_unknownstate_customer = a_customer.get_all_shelved_customers()
                     if a_unknownstate_customer: # it will return false if the customer isnt shelved so only add it if we got a customer back
                         if a_unknownstate_customer not in self.all_shelved_chat_customers:
                             self.all_shelved_chat_customers.append(a_unknownstate_customer)
+        # print(" ")
+        # print(f"All Open List => {self.all_open_chat_customers}\n")
                 
     def draw(self):
         self.screen.fill(BGCOLOR)
@@ -85,17 +92,6 @@ class Game:
         # -- create a surface for the bottom bar which will hold shelved chats --
         self.chat_bar_surf = pg.Surface((WIDTH, 60))
         self.chat_bar_surf.fill(TAN) 
-        # -- loop the all sprites group and do draw stuff -- 
-        for sprite in self.all_sprites:
-            # -- orders tab --
-            if isinstance(sprite, New_Orders_Tab):
-                if self.display_tabs[0] == "new_orders":
-                    sprite.draw()
-            # -- chats tab --
-            if isinstance(sprite, Chats_Tab):
-                if self.display_tabs[0] == "chats":
-                    sprite.draw()            
-        # -- end all sprites loop -- 
         # -- loop all the open chat customers (not in all sprites) and draw their chatboxes to the open tabs image -- 
         for index, a_customer in enumerate(self.all_open_chat_customers):
             if self.display_tabs[0] == "new_orders":   
@@ -103,11 +99,25 @@ class Game:
                     self.new_orders_tab.image = a_customer.draw_open_chatbox(index, self.new_orders_tab.image) 
             if self.display_tabs[0] == "chats":
                 if isinstance(a_customer, Customer): 
-                    self.new_orders_tab.image = a_customer.draw_open_chatbox(index, self.chats_tab.image)
+                    self.chats_tab.image = a_customer.draw_open_chatbox(index, self.chats_tab.image)
         # -- loop all the shelved chat customers and draw their chatboxes to the bottom chatbar bar --
         for index, a_customer in enumerate(self.all_shelved_chat_customers):
             if isinstance(a_customer, Customer): 
                 a_customer.draw_shelved_chatbox(index)
+        # -- loop the all sprites group and do draw stuff -- 
+        for sprite in self.all_sprites:
+            # -- orders tab --
+            # HUGE N0TE > can try use issubclass here, will handle both (this is why i love learning so much)
+            if isinstance(sprite, New_Orders_Tab):
+                if self.display_tabs[0] == "new_orders":
+                    sprite.draw()
+                    self.screen.blit(sprite.image, sprite.pos)
+            # -- chats tab --
+            if isinstance(sprite, Chats_Tab):
+                if self.display_tabs[0] == "chats":
+                    sprite.draw()            
+                    self.screen.blit(sprite.image, sprite.pos)
+        # -- end all sprites loop -- 
         # -- draw the chatbar surface across the bottom of the screen --
         self.screen.blit(self.chat_bar_surf, (0, HEIGHT - self.chat_bar_surf.get_height())) # hard code height var once display is sorted (is also in the above surface init obvs)     
         # -- finally flip the display -- 
@@ -120,7 +130,12 @@ class Game:
         for event in pg.event.get():
             if event.type == pg.QUIT:
                 self.quit()
-            # -- mouse up 
+            # -- mouse down --
+            if event.type == pg.MOUSEBUTTONDOWN: # <- own function duh
+                self.mouse_click_down = True
+            else:
+                self.mouse_click_down = False
+            # -- mouse up --
             if event.type == pg.MOUSEBUTTONUP: # <- own function duh
                 self.mouse_click_up = True
             # -- key up -- 
@@ -131,7 +146,7 @@ class Game:
                     # cycle the tabs by adding the current tab to the end of the list and popping it off the front                    
                     self.display_tabs.append(self.display_tabs[0]) 
                     self.display_tabs.pop(0)
-                    print(self.display_tabs)                    
+                    # print(self.display_tabs)                    
             # -- key down --
             if event.type == pg.KEYDOWN:
                 if event.key == pg.K_ESCAPE: