+++ v0_021/sprites.py
@@ -1,6 +1,6 @@
 import pygame as pg
 from settings import *
-from random import choice
+from random import choice, randint
 vec = pg.math.Vector2
 
 
@@ -15,6 +15,7 @@ class Customer(pg.sprite.Sprite):
         self.game_state = "active"
         self.chatbox_state = "shelved" # opened or shelved
         self.my_name = choice(["James","Jim","John","Jack","Josh","Tim","Tom","Jonathon","Steve","Carl","Mike","Brian"])
+        self.my_name += " " + choice(["A","B","C","D","E","F","G","H","I","J","K","L"])
         # add a display id - e.g KX139 or sumnt (have it be zones or sumnt but its slightly obscure so you dont twig it for a while, maybe like EWSN for cardinal directions)
         # -- chatbox specific vars --         
         self.shelved_chat_width = 200 
@@ -22,6 +23,15 @@ class Customer(pg.sprite.Sprite):
         # -- initialising other empty instance vars --
         self.chatbox_shelved_destination_rect = False
         self.chatbox_opened_destination_rect = False
+        # -- customer info --
+        self.times_purchased = 0
+        self.loyalty_member = False
+        self.last_order = False
+        self.total_spend = f"{randint(0,250)}:{randint(0,99)}"
+        self.display_location = "location"
+        self.my_order = [{"item":"nuka cola", "quantity":2},{"item":"egg fried rice", "quantity":1}]
+        # new test
+        self.force_to_mouse_position = False
 
     def get_all_shelved_customers(self):
         """ returns the given instance if it is in a shelved state, else returns false """
@@ -48,7 +58,7 @@ class Customer(pg.sprite.Sprite):
         
     def draw_open_chatbox(self, index, surf:pg.Surface): 
         # -- main open chatbox bg surf and dimensions --
-        self.opened_chat_width = self.shelved_chat_width # fine for now, will wanna update when adding buttons and interactions to the open chat window tho
+        self.opened_chat_width = 400
         self.opened_chat_height = 300 # will want this to be dynamic obvs but unsure as per the spec so just doing whatever and will refactor it all tomo
         self.chat_box_surf = pg.Surface((self.opened_chat_width, self.opened_chat_height))
         # -- alternate colours -- 
@@ -59,6 +69,11 @@ class Customer(pg.sprite.Sprite):
         # -- draw the chatters name to the shelved chatbox surf --
         chatbox_title = self.game.FONT_VETERAN_TYPEWRITER_26.render(f"{self.my_name}", True, WHITE)
         self.chat_box_surf.blit(chatbox_title, (10, 10))
+        # -- some other general info text on the character -- 
+        chatbox_customer_info = self.game.FONT_BOHEMIAN_TYPEWRITER_20.render(f"Spent: ${self.total_spend}, Location: {self.display_location}", True, WHITE)
+        chatbox_customer_chat_text = self.game.FONT_BOHEMIAN_TYPEWRITER_20.render(f"I would like a {self.my_order[0]['item']}", True, WHITE)
+        self.chat_box_surf.blit(chatbox_customer_info, (10, 40))
+        self.chat_box_surf.blit(chatbox_customer_chat_text, (10, 70))
         # -- store its pos so we can move it in future -- 
         cascading_offset = 40 # when a new one is open initially its offset is the previous position (tho u need to cap this, then increase the y at a certain point tho dont reeeeally need to do that for a tech demo)
         self.open_position = (150 + (index * cascading_offset), 150 + (index * cascading_offset)) # cascading_offset ensures they arent drawn on top of each other initially
@@ -73,8 +88,11 @@ class Customer(pg.sprite.Sprite):
         self.true_minimise_button_rect.move_ip(WIDTH - surf.get_width(), HEIGHT - surf.get_height())
         # -- final blit to the given (active) Tab surface --
         surf_copy = surf.copy()
+        # if self.force_to_mouse_position:
+            # self.open_position = pg.mouse.get_pos()
         self.chatbox_opened_destination_rect = surf_copy.blit(self.chat_box_surf, self.open_position) 
         self.chatbox_opened_destination_rect.move_ip(WIDTH - surf.get_width(), HEIGHT - surf.get_height()) # is like 20, 20 or sumnt as the surf is the tab but obvs lets still do it and still do it dynamically anyway
+        # --
         return surf_copy
 
     def check_click_shelved_chatbox(self):
@@ -83,40 +101,63 @@ class Customer(pg.sprite.Sprite):
             if self.chatbox_shelved_destination_rect:
                 if self.chatbox_shelved_destination_rect.collidepoint(pg.mouse.get_pos()):
                     self.chatbox_state = "opened"
-                    print(f"{self.my_name} => {self.chatbox_state}")
                     self.chatbox_shelved_destination_rect = False
                     self.game.all_shelved_chat_customers.remove(self)
+                    # --
                     return self
 
     def check_click_opened_chatbox(self):
         """ if the rect has been clicked we set this customers chatbot state to opened """
+        # if self.chatbox_opened_destination_rect:
+        #     if self.chatbox_opened_destination_rect.collidepoint(pg.mouse.get_pos()):
+        #         print(f"{self.my_name} : {self.chatbox_opened_destination_rect = }")
         if self.game.mouse_click_up:
             if self.chatbox_opened_destination_rect:
                 if self.chatbox_opened_destination_rect.collidepoint(pg.mouse.get_pos()):
                     if self.true_minimise_button_rect.collidepoint(pg.mouse.get_pos()):
                         # -- this does shelving --
-                        self.chatbox_state = "shelved"
-                        print(f"{self.my_name} => {self.chatbox_state}")
+                        self.chatbox_state = "shelved"                    
                         self.chatbox_opened_destination_rect = False
                         self.game.all_open_chat_customers.remove(self)
+                    else:
+                        # -- this handles bringing clicked chatbox to front -- 
+                        self.game.all_open_chat_customers.remove(self)
+                        self.game.all_open_chat_customers.append(self)
+                    # --
+                    return self                    
+    
+    # abandoned for now
+    def move_click_hold_opened_chatbox(self):
+        if self.game.mouse_click_down:
+            if self.chatbox_opened_destination_rect:
+                if self.chatbox_opened_destination_rect.collidepoint(pg.mouse.get_pos()):
+                    self.force_to_mouse_position = True
+        else:
+            self.force_to_mouse_position = False
 
-                        # -- this does bring to front -- 
-                        # self.game.all_open_chat_customers.remove(self)
-                        # self.game.all_open_chat_customers.append(self)
-                        
-                        return self                    
+        # what you would wanna do is
+        # check if force_to_mouse_pos is true
+        # if it is ur holding a rect
+        # if u click again drop the rect at the mouse pos
+        # and reset the var
+        # better just to do refactor tho imo
+
+                    
+                
+            
 
 
 
-# add button for minimise and add the shelving functionality to it
-# - may require enumerating (else its like a customer self var) to get the order and blitting the top on that way (use on hover print with a \n to figure it out)
-# then obvs add the bring to front functionality on click
-# - try to get this working in some way specific to the layering as its weird rn? 
-#   - say when you click if u click on more than one it just always takes the toppest one bosh)
+# start on da order stuff baybay!
+# - section out the areas so main area, right hand bar with order slip, order buttons, and chat windows
+# - do this little menu book thing that you can bring up with the item names and numbers, and also ig change page too
+
+# then since no camera, make the new version here with the ui stuff too
+# - this way can then use the camera to animate in and out from that position (start end level, and use can too ig - animate the little window with traffic ig, defo make it a city)
+# - and then also animate the book coming up easier
+# - that being said all of this could still be achieved here, except doing the ui will probably take just as long to add here as to refactor so just refactor lol
 
 # try get a lil bit of graphics in 🥺, see the inspo 
-
-# then make and save to a public repo, and continue
 
 # make the open chatboxes a better size with some additional light text (text box, image space, name, last reply or whatever, how long a customer for etc etc obvs just do quick now)
 # then lets get like 2 or 3 super basic order buttons down and start getting an order together!
@@ -160,7 +201,17 @@ class Browser_Tab(pg.sprite.Sprite):
         """ runs in main draw loop, draw to our background image then draw out background image to the screen every frame """
         title = self.game.FONT_VETERAN_TYPEWRITER_26.render(f"{self.my_tab_name}", True, BLUEMIDNIGHT) # FONT_VETERAN_TYPEWRITER_26 # FONT_BOHEMIAN_TYPEWRITER_26
         self.image.blit(title, (30,30))
-        self.game.screen.blit(self.image, self.pos)
+        # self.game.screen.blit(self.image, self.pos)
+        if isinstance(self, New_Orders_Tab):
+            pg.draw.rect(self.image, RED, pg.Rect(20, 20, 20, 20))
+        else:
+            pg.draw.rect(self.image, BLUE, pg.Rect(40, 40, 20, 20))
+
+    def update(self):
+        self.wipe_surface()
+
+    def wipe_surface(self):
+        self.image.fill(WHITE)
 
 
 class New_Orders_Tab(Browser_Tab):